requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coremltools==2.1.0
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
imutils==0.5.2
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
lxml==4.3.1
MarkupSafe==1.1.0
numpy==1.15.4
openapi-codec==1.3.2
opencv-python==3.4.3.18
Pillow==5.3.0
protobuf==3.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytesseract==0.2.6
python-bidi==0.4.0
pytz==2018.7
PyYAML==3.13
regex==2019.2.7
requests==2.21.0
ruamel.yaml==0.15.85
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.17.1
